third_party/blink/web_tests/platform/generic/fast/dom/getElementsByTagName-localName-matching2-expected.txt

PASS container.getElementsByTagName('linearGradient').length is 2
PASS container.getElementsByTagName('lineargradient').length is 0
PASS container.getElementsByTagName('LINEARGRADIENT').length is 0
PASS container.getElementsByTagName('FOO').length is 1
PASS container.getElementsByTagName('foo').length is 1
PASS container.getElementsByTagName('Foo').length is 0
PASS container.getElementsByTagName('img').length is 2
PASS container.getElementsByTagName('IMG').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'linearGradient').length is 1
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'lineargradient').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'LINEARGRADIENT').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'FOO').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'foo').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'Foo').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'img').length is 1
PASS container.getElementsByTagNameNS('http://www.w3.org/2000/svg', 'IMG').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'linearGradient').length is 1
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'lineargradient').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'LINEARGRADIENT').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'FOO').length is 1
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'foo').length is 1
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'Foo').length is 0
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'img').length is 1
PASS container.getElementsByTagNameNS('http://www.w3.org/1999/xhtml', 'IMG').length is 0
PASS successfullyParsed is true

TEST COMPLETE
